chore(quality): moved tide config methods
Into separate sub package
