+++ pkg/prow/config/types.go
@@ -1,0 +1,9 @@
+package config
+
+type Kind string
+
+const (
+	Application Kind = "APPLICATION"
+	Environment Kind = "ENVIRONMENT"
+	Protection  Kind = "PROTECTION"
+)
